Fix reset vector patching logic

+++ nodes/bl_core/bl_slave.h
@@ -15,12 +15,12 @@ class BLSlave : public LINDev
 public:
     BLSlave() : LINDev(true) {}
 
-    static bool     is_program_valid();
-    static void     run_program();
+    static bool         is_program_valid();
+    static void         run_program();
 
 protected:
-    virtual void    st_header_received() override;
-    virtual void    st_response_received(Response &frame) override;
+    virtual void        st_header_received() override;
+    virtual void        st_response_received(Response &frame) override;
 
 private:
     static const uint16_t kInfoPage = FLASHEND - (SPM_PAGESIZE - 1);
@@ -33,12 +33,12 @@ private:
     static uint16_t        _page_status;
     static uint16_t        _program_end;
     static uint16_t        _reset_vector;
-    static uint8_t         _page_buffer[];
+    static uint16_t        _page_buffer[];
 
-    static void     set_page_address(uint16_t address);
-    static bool     add_page_byte(uint8_t byte);
-    void            send_response();
-    static void     program_page();
-    static void     update_program_info();
-    static uint16_t get_program_crc(uint16_t length);
+    static void         set_page_address(uint16_t address);
+    static bool         add_page_byte(uint8_t byte);
+    void                send_response();
+    static void         program_page();
+    static void         update_program_info();
+    static uint16_t     get_program_crc(uint16_t length);
 };
